requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.16.0
psycopg==3.2.7
psycopg-binary==3.2.7
sqlparse==0.5.3
typing_extensions==4.13.2
